requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.4
django-environ==0.11.2
psycopg2-binary==2.9.7
sqlparse==0.4.4
typing-extensions==4.7.1
djangorestframework==3.14.0